# lib/api-zod/src/generated/types/activity.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * XmiX private photo-sharing API for organizations
 * OpenAPI spec version: 0.1.0
 */
import type { ActivityKind } from './activityKind';

export interface Activity {
  id: string;
  kind: ActivityKind;
  message: string;
  createdAt: string;
}