# lib/api-zod/src/generated/types/adminOverview.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * XmiX private photo-sharing API for organizations
 * OpenAPI spec version: 0.1.0
 */
import type { Activity } from './activity';
import type { AdminRole } from './adminRole';
import type { Institution } from './institution';

export interface AdminOverview {
  role: AdminRole;
  institution: Institution | null;
  totalInstitutions: number;
  totalUsers: number;
  activeUsers: number;
  allocatedSeats: number;
  usedSeats: number;
  utilizationPercent: number;
  recentActivity: Activity[];
}