# lib/api-zod/src/generated/types/driveUploadRequest.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * XmiX private photo-sharing API for organizations
 * OpenAPI spec version: 0.1.0
 */

export interface DriveUploadRequest {
  base64: string;
  filename: string;
  folderId: string;
  caption?: string;
}